refactor: expose page widget boundaries
