--- src/pages/canvas/features/CanvasControls/ui/CanvasControls.tsx
@@ -1,11 +1,0 @@
-import CopyAllCodeButton from './CopyAllCodeButton';
-import ResetViewButton from './ResetViewButton';
-
-export function CanvasControls() {
-  return (
-    <>
-      <ResetViewButton />
-      <CopyAllCodeButton />
-    </>
-  );
-}
