+++ src/pages/canvas/widgets/CanvasControls/ui/CanvasControls.tsx
@@ -1,0 +1,11 @@
+import CopyAllCodeButton from '../../../features/CanvasControls/ui/CopyAllCodeButton';
+import ResetViewButton from '../../../features/CanvasControls/ui/ResetViewButton';
+
+export function CanvasControls() {
+  return (
+    <>
+      <ResetViewButton />
+      <CopyAllCodeButton />
+    </>
+  );
+}
