+++ src/pages/git/widgets/GitCommit/ui/GitCommitBox.tsx
