+++ src/pages/git/widgets/GitHistory/ui/GitHistorySection.tsx
